algo calculate distance + calculate fuel

+++ TUI_SimulatorFlightCosts/Domain/TUI.SimulatorFlights.Domain/IFlight.cs
@@ -3,12 +3,12 @@
     public interface IFlight
     {
         string Name { get;}
-        void RegisterDepartureAirport(string airportName, decimal latitude, decimal longitude);
-        void RegisterDestinationAirport(string airportName, decimal latitude, decimal longitude);
+        void RegisterDepartureAirport(string airportName, double latitude, double longitude);
+        void RegisterDestinationAirport(string airportName, double latitude, double longitude);
         void Rename(string flightName);
         double CalculateDistance();
         double CalculateFuelConsumption();
         void SetFuelConsumptionPerDistancePerFlightTime(int fuelConsumptionPerDistancePerFlightTime);
-        void SetFuelConsumptionTakeoffEffort(int fuelConsumptionTakeoffEffort);
+        void SetFuelConsumptionTakeoffEffort(int fuelConsumptionTakeoffEffort);        
     }
 }
